+++ lib/src/clojure_lsp/feature/document_symbol.clj
@@ -5,8 +5,6 @@
 
 (set! *warn-on-reflection* true)
 
-(def declaration-buckets #{:namespace-definitions :var-definitions})
-
 (defn element->symbol-kind [el]
   (cond
     (#{:namespace-usages :namespace-definitions} (:bucket el)) :namespace
@@ -15,9 +13,14 @@
     (#{:var-definitions :var-usages} (:bucket el)) :variable
     :else :null))
 
-(defn ^:private element-symbol [e]
+(defn element->name [{elem-name :name :keys [dispatch-val]}]
+  (cond-> (name elem-name)
+    dispatch-val (str " " dispatch-val)))
+
+(defn ^:private element->document-symbol [e]
   (shared/assoc-some
-    {:kind (element->symbol-kind e)
+    {:name (element->name e)
+     :kind (element->symbol-kind e)
      :range (shared/->scope-range e)
      :selection-range (shared/->range e)
      :tags (cond-> []
@@ -25,17 +28,8 @@
     :detail (when (:private e)
               "private")))
 
-(defn ^:private var-defs [db filename]
-  (->> (q/find-var-definitions db filename true)
-       (map (fn [e]
-              (assoc (element-symbol e)
-                     :name (-> e :name name))))))
-
-(defn ^:private defmethods [db filename]
-  (->> (q/find-defmethods db filename)
-       (map (fn [e]
-              (assoc (element-symbol e)
-                     :name (-> e :name name))))))
+(defn ^:private symbol-order [{:keys [selection-range]}]
+  [(:line (:start selection-range)) (:character (:start selection-range))])
 
 (defn document-symbols [db filename]
   (let [namespace-definition (q/find-namespace-definition-by-filename db filename)]
@@ -46,7 +40,8 @@
       :selection-range (if namespace-definition
                          (shared/->scope-range namespace-definition)
                          shared/full-file-range)
-      :children (->> (concat (var-defs db filename)
-                             (defmethods db filename))
-                     (sort-by #(shared/position->line-column (:start (:selection-range %))))
+      :children (->> (concat (q/find-var-definitions db filename true)
+                             (q/find-defmethods db filename))
+                     (map element->document-symbol)
+                     (sort-by symbol-order)
                      vec)}]))